+++ Src/Private/Get-AbrVb365ObjectRepository.ps1
@@ -5,7 +5,7 @@ function Get-AbrVB365ObjectRepository {
     .DESCRIPTION
         Documents the configuration of Veeam VB365 in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.3.8
+        Version:        0.3.11
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -19,14 +19,14 @@ function Get-AbrVB365ObjectRepository {
     )
 
     begin {
-        Write-PScriboMessage "Repository InfoLevel set at $($InfoLevel.Infrastructure.Repository)."
+        Write-PScriboMessage -Message "Repository InfoLevel set at $($InfoLevel.Infrastructure.Repository)."
     }
 
     process {
         try {
             $script:ObjectRepositories = Get-VBOObjectStorageRepository -WarningAction SilentlyContinue | Sort-Object -Property Name
             if (($InfoLevel.Infrastructure.Repository -gt 0) -and ($ObjectRepositories)) {
-                Write-PScriboMessage "Collecting Veeam VB365 Object Repository."
+                Write-PScriboMessage -Message "Collecting Veeam VB365 Object Repository."
                 Section -Style Heading2 'Object Repositories' {
                     $ObjectRepositoryInfo = @()
                     foreach ($ObjectRepository in $ObjectRepositories) {
@@ -107,7 +107,7 @@ function Get-AbrVB365ObjectRepository {
                 }
             }
         } catch {
-            Write-PScriboMessage -IsWarning "Object Repository Section: $($_.Exception.Message)"
+            Write-PScriboMessage -IsWarning -Message "Object Repository Section: $($_.Exception.Message)"
         }
     }